Oblivion-main/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oblivion-redteam"
version = "0.1.0"
description = "LLM-assisted adversary emulation framework"
readme = "readme.md"
license = {text = "MIT"}
authors = [{name = "Oblivion Team", email = "security@example.com"}]
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.7",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "metasploit-py>=1.0.0",
    "rich>=13.7.0",
    "python-nmap>=0.7.1",
    "tenacity>=8.2.3",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.24.0"
]

[project.scripts]
oblivion = "ob1.__main__:cli"